Small type rearrangement
